modified : libft.h

+++ srcs/ft_strdup.c
@@ -6,14 +6,16 @@
 /*   By: mnchabel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 11:35:39 by mnchabel          #+#    #+#             */
-/*   Updated: 2019/05/27 13:42:12 by mnchabel         ###   ########.fr       */
+/*   Updated: 2019/05/27 15:03:35 by mnchabel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "../headers/libft.h"
 
 char	*ft_strdup(const char *s)
 {
 	char	*dest;
-	int		i;
+	size_t	i;
 
 	i = 1;
 	while (s[i])
